requirements.txt
Faker==40.1.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
Unidecode==1.4.0
